app: implement Post component
since we are using TSX and not JSX, some of the functions of
react-router-dom i.e. Form, useActionData, useSubmit are very
hard to use in terms of readability and accessibility. there is
a possibility that if we go down that rabbit-hole, it would
be lots of boilerplate to verify the types that typescript wants